refactor: move modality enums to enums.ts
