+++ src/lib/enums.ts
@@ -1,6 +1,7 @@
 import { z } from "zod";
 
 export const DaysEnumSchema = z.enum(["M", "T", "W", "H", "F", "S"]);
+
 export type DaysEnum = z.infer<typeof DaysEnumSchema>;
 
 export const ColorsEnumSchema = z.enum([
@@ -24,3 +25,13 @@ export const ColorsEnumSchema = z.enum([
 ]);
 
 export type ColorsEnum = z.infer<typeof ColorsEnumSchema>;
+
+export const ModalityEnumSchema = z.enum([
+  "HYBRID",
+  "F2F",
+  "ONLINE",
+  "PREDOMINANTLY ONLINE",
+  "TENTATIVE",
+]);
+
+export type ModalityEnum = z.infer<typeof ModalityEnumSchema>;
